[OkHttp] Refactor
