+++ core/src/main/scala/sttp/client/ws/internal/FutureAsyncQueue.scala
@@ -1,12 +1,12 @@
-package sttp.client.okhttp
+package sttp.client.ws.internal
 
 import java.util.concurrent.{ArrayBlockingQueue, BlockingQueue, LinkedBlockingQueue}
 
-import sttp.client.Identity
-import sttp.client.ws.internal.AsyncQueue
 import sttp.model.ws.WebSocketBufferFull
 
-class SyncQueue[T](capacity: Option[Int]) extends AsyncQueue[Identity, T] {
+import scala.concurrent.{ExecutionContext, Future, blocking}
+
+class FutureAsyncQueue[T](capacity: Option[Int])(implicit ec: ExecutionContext) extends AsyncQueue[Future, T] {
 
   private val queue: BlockingQueue[T] = capacity match {
     case Some(value) => new ArrayBlockingQueue[T](value)
@@ -25,5 +25,5 @@ class SyncQueue[T](capacity: Option[Int]) extends AsyncQueue[Identity, T] {
   /**
     * Takes an element from the queue or suspends, until one is available. May be eager or lazy, depending on `F`.
     */
-  override def poll: Identity[T] = queue.take()
+  override def poll: Future[T] = Future(blocking(queue.take()))
 }
